pyproject.toml
[project]
name = "langchain-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-deepseek>=0.1.4",
    "langchain-ollama>=0.3.8",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
]
